Improve pre-trigger sampling on below 500kHz sample rates.
The sample loop was not padded properly in the loop waiting for the
trigger to fire. As a result it was sampling at a much higher rate than
the post trigger sample rate.  I've added some delays and padded it out
a bit, it needs further measurement, but is usable now.