Update device profile clockspeed & portdelay.
Set clock speed to 16MHz.  Hopefully this value works fine with the
client since the previous 100MHz was not the actual clockspeed anyway.
Also, adjust the device.open.portdelay to 1500ms.  People can test
their own device for the fastest value, but this should be a safer
default for most people and hopefully eliminates som frustration for
first time users.