Increase portdelay so that Arduino Uno R3 works.
For some reason the Arduino Uno R3 (but not my earlier / original Uno)
needs a longer delay after reset.  I haven't investigated the cause
yet, but increasing device.open.portdelay to > 1700ms seems to fix it.
Bump to 2000 just to be safe in all cases.  This should fix the Uno R3
issues that have been reported a couple of times now.