Add additional diagnostic commands.
You can connect via the serial monitor and send ‘?’ to get a list of
commands available.  This allows some diagnostics when the board isn’t
working with the OLS client.
Sending ‘4’ does a 4MHz capture and ‘2’ dumbs the data buffer for
example.  Recommended to use this with the simulated signal via a
timer.  See the bottom of the setup() function and change the ‘#if 0’
to ‘#if 1’ to enable it.
Also, fix a long standing bug that broke 2MHz captures and the
ATmega168. (improper #if defined check)